requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.2

# AI/ML
openai==1.10.0

# WhatsApp Integration
twilio==8.11.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Email
aiosmtplib==3.0.1

# Document Processing
pypdf2==3.0.1
python-docx
python-statemachine>=2.0.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
mypy==1.8.0
